+++ tests/testthat.R
@@ -6,6 +6,7 @@
 # * https://r-pkgs.org/testing-design.html#sec-tests-files-overview
 # * https://testthat.r-lib.org/articles/special-files.html
 
+Sys.setenv(RGL_USE_NULL = "true")
 library(testthat)
 library(D3mirt)